requirements.txt
pyyaml>=6.0
colorama>=0.4.6
tqdm>=4.65.0
python-nmap>=0.7.1
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.3